Move instantiation of objects which depend on _helixResourceManager to after invocation of PinotHelixResourceManager::start (#3713)
Certain objects which are instantiated in the ControllerStarter need an instance of PinotHelixResourceManager. Even though PinotHelixResourceManager is instantiated in the ControllerStarter constructor right at the beginning, it cannot be used until PinotHelixResourceManager::start is invoked. The start is invoked in ControllerStarter::start(). Every object that depends on PinotHelixResourceManager should be created after this invocation of start
This PR moves all the dependents to after PinotHelixResourceManager being ready